Use code from author to initialize sim.
