+++ packages/agent-simulation/src/components/agent-simulation.scss
@@ -5,7 +5,7 @@
   gap: 10px;
   flex-direction: column;
 
-  .blocklyCode {
+  .code {
     background-color: #eee;
     border: 1px solid #ccc;
     font-family: monospace;
