+++ packages/agent-simulation/src/components/types.ts
@@ -4,27 +4,30 @@ import { IAuthoringInteractiveMetadata, IRuntimeInteractiveMetadata } from "@con
 // TS interfaces are not available in runtime in contrast to JSON schema.
 
 export interface IAuthoredState extends IAuthoringInteractiveMetadata {
-  version: number;
-  hint?: string;
+  code: string;
   dataSourceInteractive?: string;
+  hint?: string;
+  version: number;
 }
 
 export interface IInteractiveState extends IRuntimeInteractiveMetadata {
-  submitted?: boolean;
   exampleInteractiveState?: string;
+  submitted?: boolean;
 }
 
 export const DefaultAuthoredState: Omit<Required<IAuthoredState>, "questionSubType"|"required"|"prompt"> = {
-  version: 1,
-  questionType: "iframe_interactive",
-  hint: "",
+  code: "",
   dataSourceInteractive: "",
+  hint: "",
+  questionType: "iframe_interactive",
+  version: 1,
 };
 
 export const DemoAuthoredState: IAuthoredState = {
-  version: 1,
-  questionType: "iframe_interactive",
-  prompt: "<p>TODO: Add the final real prompt in types.ts</p>",
-  hint: "TODO: Add the final real hint in types.ts",
+  code: "",
   dataSourceInteractive: "",
+  hint: "TODO: Add the final real hint in types.ts",
+  prompt: "<p>TODO: Add the final real prompt in types.ts</p>",
+  questionType: "iframe_interactive",
+  version: 1,
 };
